prettified

+++ sctp_proxy/project_folder/sctp_primitives_client.h
@@ -39,8 +39,8 @@
  * \param length Length of message received
  * @return Execution result
  */
-typedef int (*client_sctp_recv_callback)(uint32_t stream,
-                                  uint8_t *buffer, uint32_t length);
+typedef int (*client_sctp_recv_callback)(uint16_t stream, uint8_t *buffer,
+                                         uint32_t length);
 
 /** \brief SCTP connected callback prototype. Will be called once the
  * association is ready.
@@ -73,8 +73,9 @@ int sctp_connect_to_remote_host(char *local_ip_addr[], int nb_local_addr,
  * \param length Buffer length
  * @return < 0 in case of failure
  */
-int client_sctp_send_msg(sctp_data_t *sctp_data_p, uint16_t ppid, uint16_t stream,
-                  const uint8_t *buffer, const uint32_t length);
+int client_sctp_send_msg(sctp_data_t *sctp_data_p, uint16_t ppid,
+                         uint16_t stream, const uint8_t *buffer,
+                         const uint32_t length);
 
 /** \brief Flush the FIFO of messages from the socket
  * \param sctp_data_p
